Switch to provided jsr305 and nullability improvements (#69)
This takes a feather out of Square's recent blog post and brings back the jsr305 dependency as a provided dependency. Benefits are static analysis and kotlin code still can leverage this.

Main functional differences are annotating a `package-info` with `@ParametersAreNonnullByDefault` annotation and switching to javax' `@Nullable` annotation instead of RxJava's.