Structured Android components (and some refactorings) (#111)
* Move android components to /android directory

* Opportunistic update to Gradle 4.2.1

* Extract TestAndroidLifecycleScopeProvider to own module, rename

This renames TestAndroidLifecycleScopeProvider to the more appropriate TestLifecycleOwner. Amongst other things, this makes the extensions/common dependencies only part of that artifact and not a part of the mainline artifact

* Opportunistic update kotlin plugin name to newer qualified one

* Switch to static creators for TestLifecycleOwner

* Add autodispose-android-kotlin artifact

* Add autodispose-android-archcomponents-kotlin artifact

* Add autodispose-android-archcomponents-test-kotlin artifact

* arch components rc1

* Build tools 26.0.2

* Add consumer proguard rules for javaxextras

Resolves #112

* Common 1.0.3 + include in archcomponents

This is apparently necessary for generated code

* Make the test use something other than .test for name to avoid conflicts

* Remove manifest, set in defaultConfig instead

* Ignore restricted API warning in generated adapter code

* Add generated adapter to proguard rules in case

* Add convenience Function<Lifecycle.Event, Lifecycle.Event> resolver API

This also simplifies the multiple constructors we had going on

* Update stop check to match slinky behavior

Due to a behavior change, stop events just mean it falls back to "created", which presumably could restart. To match this, I've changed the test to measure this such that if the last even was pause, we don't assume destroy, and actually just stop in onStop like normal

* Remove backfilling from TestLifecycleOwner, as registry handles it now