Incorporate feedback from #130 (#138)
* Remove unnecessary synchronized blocks

* Fix introspection typo

* Fix introspection typo in observer

* Make AutoDispose ScopeHandlers final

* Extract DisposableMaybeObserver and setOnce first

* Opportunistic tweak of RxErrorsRule

Adds a convenience assertion and uses peek instead of poll for hasError

* Lazyset disposables in scope callbacks to prevent double subscriptions

These were missing before and necessary to mark this observer as appropriately disposed after, though without calling dispose()

* Inline lazyDispose() methods

Saving dex methods

* Wire up HalfSerializer to observer and subscribers

I'm not 100% I've done this right, but leveraging some existing examples in RxJava as my guide. I've copied in the relevant internal rxjava APIs (since they're internal) and also adjusted the onNext methods to return true or false indicating whether or not they terminated the underlying observer. Will seek further feedback during the PR. Main thing I'm not sure of is if it does terminate, do we need to dispose the lifecycle or does this mean that onError or onComplete have been hit in the main observer already (and doing so is a double dispose condition)

* Fix ViewAttachEventsObservable not observing observer contract

* Fix LifecycleEventsObservable not observing observer contract

* Inline lazyDispose() impl

* Remove unnecessary nullcheck in request()

* Suppress nullaway false positive

* Initialize atomic throwables inline

* Use deferred requesting

* Call delegate.onSubscribe() first, remove callSubscribeIfNecessary

* Check if observer is disposed in lifecycle events

* Check if listener is disposed in viewattacheventsobservable

* Remove private to avoid synthetic accessor

* Use non-rx-internal subscriptionhelper APIs

* Opportunistic fix missing synchronized in overridden method

* Remove callSubscribeIfNecessary, avoid duplicate subscribes

* Switch orders of lazySet + dispose to flush

* Optimize lifecycle error propagation a bit

We were disposing the lifecycle a second time after, and unnecessarily disposing the main disposable twice too