Migrate ObservableScoper to delegate to as() APIs + some cleanup (#189)
* Remove superfluous exception signature

* Rename Scoper to BaseAutoDisposeConverter

This class was always package private and not public API

* Rename ParallelFlowableScoper to AutoDisposeParallelFlowableConverter

More conventional name we're using going forward

* Create new AutoDisposeObservableConverter

Matching the ParallelFlowable implementation

* Use AutoDisposableObservableConverter in AutoDispose.java

* Migrate ObservableScoper to delegate to the as() APIs

This migrates ObservableScoper to delegate to the as() APIs. This was a bit weird because the scoper APIs were a bit more flexible (due to the nature of to()) with generics, and we have to cover that with an identity function. This makes them slightly more overhead now, but we're ok with that because they're a deprecated API that we're encouraging people to migrate off of.

* Make identityFunctionForGenerics() package (more) private

Then we don't need to worry about people subclassing a scoper and seeing this either

* Checkstyle fixes

* Remove dependency of converters on BaseAutoDisposeScoper

They actually don't need it!

* Inline AutoDisposeObservableConverter

This reduces allocations by 1 and also the size of the library a bit while maintaining overall less state for what are effectively synchronous operations

* Inline AutoDisposeParallelFlowableConverter

This reduces allocations by 1 and also the size of the library a bit while maintaining overall less state for what are effectively synchronous operations