Lint Module (#282)
**Description**: Adds a lint module that checks for AutoDispose usage and raises errors when it's not being used to manage your subscriptions. 

The lint checks mainly takes advantage of the fact that when subscription is managed through AutoDispose, the `as()` operator returns an `ObservableConverter` instead of a plain old `Observable`. So we check for instances of `subscribe()` on all the five reactive types and raise an error. Let me know if that logic holds up. 

There is also a good scope for suggesting quick fixes. For example, we can find when the `subscribe()` is in an Activity/Fragment and then suggest quick fixes accordingly. I'll add that in a follow up so as to not introduce noise here. 

**Related issue(s)**: #203