Make error-prone checker add custom scopes to default scopes (#308)
This adds a new flag `isScopeExclusive` (open to suggestions) that will make the classes supplied in `TypesWithScope` exclusive and will only perform the check on those classes. 

In the absence of that flag, we will *add* the custom scope classes to the `DEFAULT_CLASSES_WITH_LIFECYCLE` instead of ignoring the default ones. 

Part of #298