Add animal-sniffer to guard against Java 8 only APIs (#336)
I was looking at integrating Animal Sniffer elsewhere and I found that the integration in AutoDispose was non-existent. In #264, I added a Java 8 API to show that Animal Sniffer works. Then when I was supposed to revert the Java 8 API commit, I reverted actual animal sniffer integration. 

![source](https://user-images.githubusercontent.com/2227639/54553592-c792e080-496f-11e9-8d71-a8cbb06e79eb.gif)

So now, this build will fail because the Java 8 API is still existent (it didn't harm anyone since it was only a test that wasn't used by anyone). I'll remove it once this build fails on CI.